chore: Update website link styling in Website component
The website link styling in the Website component has been updated to improve the user experience. The font weight has been changed to "semibold" and the hover effect has been modified to change the text color to white and the background color to black. This change ensures that the link stands out and is more visually appealing to users.