Update mobile page with information about downloading Android APKs
The mobile page has been updated to include information about downloading Android APKs. This change adds a new section with text explaining that the listed APKs can only be downloaded on mobile devices and are specifically for Android. It also mentions that an iOS APK version will be posted soon. This update provides clarity to users about the availability of the APKs and the platforms they are compatible with.